pyproject.toml
[tool.poetry]
name = "tg-bot"
version = "0.1.0"
description = "Simple tg bot"
authors = ["Zmiy3012 <semenzmiy@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
openai = ">=1.68.2,<2.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
python-telegram-bot = ">=22.0,<23.0"
pydantic-settings = ">=2.8.1,<3.0.0"
aiofiles = ">=24.1.0,<25.0.0"
aiosqlite = ">=0.21.0,<0.22.0"
tabulate = ">=0.9.0,<0.10.0"

poetry = "2.1.4"
poetry-core = "2.1.3"
pydantic = "2.11.7"
pydantic_core = "2.33.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"